Migrate to module
