+++ package.json
@@ -1,5 +1,6 @@
 {
     "name": "@openaddresses/lib",
+    "type": "module",
     "version": "2.8.0",
     "main": "oa.js",
     "repository": "git@github.com:openaddresses/lib",
@@ -9,13 +10,13 @@
         "lint": "eslint *.js src/**.js"
     },
     "bin": {
-        "oa": "oa.js"
+        "oa": "cli.js"
     },
     "dependencies": {
         "asciichart": "^1.5.25",
         "inquirer": "^8.0.0",
         "minimist": "^1.2.5",
-        "request": "^2.88.2"
+        "node-fetch": "^3.2.3"
     },
     "devDependencies": {
         "eslint": "^8.2.0",
